requirements.txt
django
pytest
djangorestframework
PyYaml
channels
asgi_redis